Refactor changes in previous commit
